Move javaify() to utils.js
